Added initial data file for mapping, rename file
